Big Changes!
Nearly all the changes are to deal with issues with terminal notifier
and cocoadialog when called from a launchdaemon. Have to spawn these as
current user or they don’t show up! Other things are config like icon
size settings.

This is the first version I am utterly confident about using in
production.